requirements.txt
alembic==1.12.0
bcrypt==4.0.1
blinker==1.6.2
branca==0.6.0
certifi==2023.7.22
charset-normalizer==3.3.1
click==8.1.7
dnspython==2.3.0
email-validator==2.0.0.post2
Flask==2.2.5
Flask-Admin==1.6.1
Flask-Bcrypt==1.0.1
Flask-Login==0.6.2
Flask-Mail==0.9.1
Flask-Migrate==4.0.5
Flask-SQLAlchemy==2.5.1
Flask-WTF==1.1.1
folium==0.14.0
ghp-import==2.1.0
greenlet==2.0.2
gunicorn==19.9.0
htmlmin==0.1.12
idna==3.4
importlib-metadata==6.7.0
itsdangerous==2.1.2
Jinja2==3.1.2
jsmin==3.0.1
Mako==1.2.4
Markdown==3.4.4
MarkupSafe==2.1.3
mergedeep==1.3.4
mkdocs==1.5.3
mkdocs-material==5.5.14
mkdocs-material-extensions==1.2
mkdocs-minify-plugin==0.3.0
numpy==1.21.6
packaging==23.2
pathspec==0.11.2
Pillow==9.5.0
platformdirs==3.11.0
psycopg2==2.9.5
Pygments==2.16.1
PyJWT==2.8.0
pymdown-extensions==10.2.1
python-dateutil==2.8.2
PyYAML==6.0.1
pyyaml_env_tag==0.1
requests==2.31.0
six==1.16.0
SQLAlchemy==1.3.19
typing_extensions==4.7.1
urllib3==2.0.7
watchdog==3.0.0
websocket-client==1.6.1
Werkzeug==2.2.3
WTForms==3.0.1
zipp==3.15.0
